requirements.txt
discord
pytz
requests
python-dotenv
